feat: scroll behaviour improved
when the machine is not running, you may are inspecting the state of the trail
so the scroll is not triggered when user scrolls horizontally.

If the automated process stoped. You need to recover the track of the latest assignment
so there is bus emitter now that inform when ever the machine started on automode